% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_api_key.R
\name{dg_api_key}
\alias{dg_api_key}
\title{Fetch the DataGolf API key from environment variables.}
\usage{
dg_api_key()
}
\value{
A character string containing the API key.
}
\description{
This function retrieves the API key set in the environment variables.
}
