% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_approach_skill.R
\name{dg_approach}
\alias{dg_approach}
\title{Fetch Golfer Approach Skill Breakdown from DataGolf.}
\usage{
dg_approach(period = "l24")
}
\arguments{
\item{period}{(default = "l24, last 24 months) other param options l12 (last 12 months) ytd (year to date).}
}
\value{
A dataframe containing golfers and their approach skill in the specified period.
}
\description{
Fetch Golfer Approach Skill Breakdown from DataGolf.
}
