% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_event_name.R
\name{dg_event_name}
\alias{dg_event_name}
\title{Fetch the event name of a current tournament from DataGolf.}
\usage{
dg_event_name(tour = "pga")
}
\value{
Event name string object from current event.
}
\description{
Fetch the event name of a current tournament from DataGolf.
}
