% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_field_updates.R
\name{dg_field_updates}
\alias{dg_field_updates}
\title{Fetch Field Updates from DataGolf.}
\usage{
dg_field_updates(tour = "pga")
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}
}
\value{
A dataframe containing field updates for the specified tour.
}
\description{
Fetch Field Updates from DataGolf.
}
