% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_historical_events.R
\name{dg_historical_events}
\alias{dg_historical_events}
\title{Fetch Historical Event List DataGolf.}
\usage{
dg_historical_events()
}
\value{
A dataframe of all historical events.
}
\description{
Fetch Historical Event List DataGolf.
}
