% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_historical_predictions.R
\name{dg_historical_predictions}
\alias{dg_historical_predictions}
\title{Fetch Historical Predictions from DataGolf.}
\usage{
dg_historical_predictions(
  event_id = "14",
  year = "2022",
  odds_format = "percent",
  format = "json"
)
}
\arguments{
\item{event_id}{string of event id (default is "14") The Masters.}

\item{year}{string of year (default is "2022").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}
}
\value{
A dataframe of raw data from a historical event prediction.
}
\description{
Fetch Historical Predictions from DataGolf.
}
