% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_historical_raw_data.R
\name{dg_historical_raw_data}
\alias{dg_historical_raw_data}
\title{Fetch Historical Raw Data for Tournament from DataGolf.}
\usage{
dg_historical_raw_data(year = "2022", event_id = "14", tour = "pga")
}
\arguments{
\item{year}{A character string specifying the year of the tournament (default is "2022").}

\item{event_id}{A character string specifying the event id of the tournament (default is "14") Scottie Scheffler Masters.}

\item{tour}{A character string specifying the tour name (default is "pga").}
}
\value{
A dataframe containing raw data from the tournament with nested objects of scores, then nested rounds.
}
\description{
Fetch Historical Raw Data for Tournament from DataGolf.
}
