% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_live_hole_scoring.R
\name{dg_live_hole_scoring}
\alias{dg_live_hole_scoring}
\title{Fetch Live Hole Scoring for Current Tournament from DataGolf.}
\usage{
dg_live_hole_scoring(tour = "pga", odds_format = "percent")
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}
}
\value{
A dataframe containing raw data from the tournament with nested objects.
}
\description{
Fetch Live Hole Scoring for Current Tournament from DataGolf.
}
