% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_live_predictions.R
\name{dg_live_predictions}
\alias{dg_live_predictions}
\title{Fetch DataGolf live tournament predictions for a specific tour.}
\usage{
dg_live_predictions(tour = "pga", odds_format = "percent", format = "json")
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}

\item{format}{A character string specifying the file format (default is "json").}
}
\value{
A dataframe containing predictions for a live event.
}
\description{
Fetch DataGolf live tournament predictions for a specific tour.
}
