% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_live_stats.R
\name{dg_live_stats}
\alias{dg_live_stats}
\title{Fetch Live Stats of Current Tournament from DataGolf.}
\usage{
dg_live_stats(
  stats = c("sg_ott", "distance", "accuracy", "sg_app", "gir", "prox_fw", "sg_putt",
    "scrambling"),
  round = "event_avg",
  tour = "pga",
  display = "value",
  file_format = "json"
)
}
\arguments{
\item{stats}{A vector string of stats with (default is "sg_ott", "distance", "accuracy", "sg_app", "gir", "prox_fw", "sg_putt", "scrambling").}

\item{round}{A character string specifying if single round or event average (default is "event_avg"), options are "event_avg", "1", "2", "3", "4".}

\item{tour}{A character string specifying the tour name (default is "pga").}

\item{display}{A character string specifying how stats are displayed if value or rank (default is "value").}

\item{file_format}{A character string specifying the file format (default is "json").}
}
\value{
A dataframe containing betting market matchups for current week tour tournament with a set market of full tournament, round, or threeball.
}
\description{
Fetch Live Stats of Current Tournament from DataGolf.
}
