% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_live_status.R
\name{dg_live_status}
\alias{dg_live_status}
\title{Fetch Live Status Info from DataGolf.}
\usage{
dg_live_status(tour = "pga", odds_format = "percent")
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}
}
\value{
A dataframe containing status info of current tournament for given tour.
}
\description{
Fetch Live Status Info from DataGolf.
}
