% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_matchups.R
\name{dg_matchups}
\alias{dg_matchups}
\title{Fetch Field Updates from DataGolf.}
\usage{
dg_matchups(
  tour = "pga",
  market = "tournament_matchups",
  odds_format = "percent"
)
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}

\item{market}{A character string specifying the market (default is "tournament_matchups"), available params ("tournament_matchups", "round_matchups", "3_balls").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}
}
\value{
A dataframe containing betting market matchups for current week tour tournament with a set market of full tournament, round, or threeball.
}
\description{
Fetch Field Updates from DataGolf.
}
