% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_outrights.R
\name{dg_outrights}
\alias{dg_outrights}
\title{Fetches betting market matchups for current week tour tournament with a set market for win, top 5, top 10, top 20, miss cut, make cut, and first round leader.}
\usage{
dg_outrights(
  tour = "pga",
  market = "win",
  odds_format = "percent",
  file_format = "json"
)
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}

\item{market}{A character string specifying the market (default is "win").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}

\item{file_format}{A character string specifying the file format (default is "json").}
}
\value{
A dataframe containing betting market matchups for current week tour tournament with a set market of full tournament, round, or threeball.
}
\description{
Fetches betting market matchups for current week tour tournament with a set market for win, top 5, top 10, top 20, miss cut, make cut, and first round leader.
}
