% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_player_decomp.R
\name{dg_player_decomp}
\alias{dg_player_decomp}
\title{Fetch Field Updates from DataGolf.}
\usage{
dg_player_decomp(tour = "pga", odds_format = "percent", format = "json")
}
\arguments{
\item{tour}{A character string specifying the tour name (default is "pga").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}

\item{format}{A character string specifying the format (default is "json").}
}
\value{
A dataframe containing field updates for the specified tour.
}
\description{
Fetch Field Updates from DataGolf.
}
