% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_player_list.R
\name{dg_player_list}
\alias{dg_player_list}
\title{Fetch the Player List from DataGolf.}
\usage{
dg_player_list()
}
\value{
A dataframe containing the list of players.
}
\description{
Fetch the Player List from DataGolf.
}
