% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_pre_tournament_predictions.R
\name{dg_pre_tournament_predictions}
\alias{dg_pre_tournament_predictions}
\title{Fetch Historical Predictions from DataGolf.}
\usage{
dg_pre_tournament_predictions(tour = "pga", odds_format = "percent")
}
\arguments{
\item{tour}{string of tour to get back current week predictions for before tournament start (default is "pga").}

\item{odds_format}{A character string specifying the odds format (default is "percent").}
}
\value{
A dataframe of event prediction.
}
\description{
Fetch Historical Predictions from DataGolf.
}
