% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_rankings.R
\name{dg_rankings}
\alias{dg_rankings}
\title{Fetch Golfer Rankings from DataGolf.}
\usage{
dg_rankings()
}
\value{
A dataframe containing field updates for the specified tour.
}
\description{
Fetch Golfer Rankings from DataGolf.
}
