% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_schedule.R
\name{dg_schedule}
\alias{dg_schedule}
\title{Fetch the Tour Schedule from DataGolf.}
\usage{
dg_schedule(tour = "pga")
}
\arguments{
\item{tour}{A character string specifying the tour name.}
}
\value{
A dataframe containing the schedule for the specified tour.
}
\description{
Fetch the Tour Schedule from DataGolf.
}
