% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dg_skill_ratings.R
\name{dg_skill_ratings}
\alias{dg_skill_ratings}
\title{Fetch Golfer Skill Ratings from DataGolf.}
\usage{
dg_skill_ratings()
}
\value{
A dataframe containing golfers skill ratings for each golfer with id as dg_id.
}
\description{
Fetch Golfer Skill Ratings from DataGolf.
}
