% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{GenerateData}
\alias{GenerateData}
\title{Generates test data frame of the students with unique ID (uppercase letters)}
\usage{
GenerateData(num_students)
}
\arguments{
\item{num_students}{a positive integer.}
}
\value{
a data frame
}
\description{
Generates test data frame of the students with unique ID (uppercase letters)
}
\examples{
# five students
GenerateData(5)
}
