% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGroups.R
\name{MakeGroups}
\alias{MakeGroups}
\title{Assign every student into groups with set number of students per group}
\usage{
MakeGroups(data, students_per_group, iterations, initial_matrix)
}
\arguments{
\item{data}{a data frame}

\item{students_per_group}{a positive integer}

\item{iterations}{a positive integer}

\item{initial_matrix}{matrix made with \code{initmat()}}
}
\value{
data frame
}
\description{
Assign every student into groups with set number of students per group,
with set number of iterations (overlap not considered)
}
