% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_to_odds.R
\name{prob_to_odds}
\alias{prob_to_odds}
\title{Converts a Probability to American Odds.}
\usage{
prob_to_odds(probability)
}
\arguments{
\item{probability}{a probability value.}
}
\value{
the american odds for that probability.
}
\description{
Converts a Probability to American Odds.
}
