% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeGroups.R
\name{MakeGroups}
\alias{MakeGroups}
\title{Assign every student into groups with set number of students per group}
\usage{
MakeGroups(data, students_per_group, iterations, initial_matrix)
}
\arguments{
\item{data}{a data frame}

\item{students_per_group}{a positive integer}

\item{iterations}{a positive integer}

\item{initial_matrix}{matrix generated with `initmat()`}
}
\value{
data frame
}
\description{
Assign every student into groups with set number of students per group,
 with set number of iterations (overlap not considered)
}
\examples{
# Assign 9 students into 3 groups of 3, with 3 iterations
data <- GenerateData(9)
M <- initmat(data$Student)
MakeGroups(data, 3, 3, M)
}
